pyproject.toml
[project]
name = "FURLONG"
version = "0.1.0"
description = "AI-driven horse racing prediction system."
authors = [
    { name = "maoki", email = "michihito.aoki@gmail.com" }
]
dependencies = [
    "flask>=3.0.3",
    "matplotlib>=3.9.0",
    "pandas>=2.2.2",
    "polars>=0.20.31",
    "plotly>=5.22.0",
    "pyg-lib==0.4.0+pt23cu121",
    "python-dotenv>=1.0.1",
    "scikit-learn>=1.5.0",
    "scipy>=1.13.1",
    "seaborn>=0.13.2",
    "statsmodels>=0.14.2",
    "torch==2.3.1+cu121",
    "torch-cluster==1.6.3+pt23cu121",
    "torch-geometric>=2.5.3",
    "torch-scatter==2.1.2+pt23cu121",
    "torch-sparse==0.6.18+pt23cu121",
    "torch-spline-conv==1.2.2+pt23cu121",
    "torchaudio==2.3.1+cu121",
    "torchvision==0.18.1+cu121",
    "lightgbm>=4.4.0",
    "pyvis>=0.3.2",
    "pyarrow>=16.1.0",
    "openpyxl>=3.1.4",
    "matplotlib-fontja>=1.0.0",
    "optuna==3.5",
    "shap>=0.45.1",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = [
    "pip>=24.1",
    "ruff>=0.4.10",
    "spyder-kernels==2.5.*",
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/furlong"]

[[tool.rye.sources]]
name = "torch"
url = "https://download.pytorch.org/whl/cu121"
type = "index"

[[tool.rye.sources]]
name = "pytorch-geometric"
url = "https://data.pyg.org/whl/torch-2.3.0+cu121.html"
type = "find-links"
